UGENE-7839 update
