+++ src/corelibs/U2Algorithm/src/misc/EnzymeModel.h
@@ -124,6 +124,7 @@ public:
      */
     int getFullLength() const;
 
+    QString generateEnzymeTooltip() const;
 };
 
 typedef QSharedDataPointer<EnzymeData> SEnzymeData;
